backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-redis==6.0.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
redis==7.1.0
sqlparse==0.5.5
python-dotenv
Pillow